+++ quick/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 namui = { path = "../namui" }
-namui-prebuilt = { path = "../namui-prebuilt" }
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.13"
