+++ quick/src/consts.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+pub const LINE_ITEM_GAP: Px = px(4.0);
+
+pub const LINE_HEIGHT: Px = px(16.0);
+pub const LINE_GAP: Px = px(4.0);
+
+pub const BLOCK_OUTER_TOP_MARGIN: Px = px(12.0);
+pub const BLOCK_OUTER_BOTTOM_MARGIN: Px = px(4.0);
+pub const BLOCK_OUTER_LEFT_RIGHT_MARGIN: Px = px(8.0);
+pub const BLOCK_INNER_MARGIN: Px = px(8.0);
