+++ quick/src/lib.rs
@@ -1,9 +1,15 @@
 mod block;
 pub mod color;
+mod consts;
+mod group;
+mod line;
 mod link;
 mod typography;
 
 pub use block::*;
+pub use consts::*;
+pub use group::*;
+pub use line::*;
 pub use link::*;
 use namui::prelude::*;
